pyproject.toml
[project]
name = "pypasswizard"
version = "0.3.2"
description = "A CLI application for managing passwords securely."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "fastparquet>=2024.11.0",
    "funkybob>=2023.12.0",
    "mypy>=1.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pandera>=0.23.1",
    "pyarrow>=20.0.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
    "setuptools>=80.0.0",
    "tabulate>=0.9.0",
    "toml>=0.10.2",
    "twine>=6.1.0",
]
